weather-bot/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
pytz>=2024.1
kalshi_python_sync>=3.10.0
certifi>=2024.0.0
